better directory structure reduce merge conflicts
